[computed_goto] Updated generated code for the VM engine, along with changes to the interpreter.
The entire VM engine routine is included in `interp_body.c`, and some of the
previous support code has been moved to `interpreter.c`. For now, I've
rewritten it to only use one engine, which is passed the block size to
use  (i.e., instead of `vm_engine_1`, `vm_engine_2`, and `vm_engine_rest`,
we just use `vm_engine_rest`). This does not seem to have much speed impact
(the compiler probably wasn't using the static size of the loops to any
great effect), and is faster to compile.

An attribute `vm_model` is added to the interpreter module that contains
the opcode-dispatch method used.

Computed gotos are used if the preprocessor symbol USE_COMPUTED_GOTO is
defined (I do it on the command line, like `EXTRA_CFLAGS=-DUSE_COMPUTED_GOTO python setup.py build`).

This experiment is probably something of a failure: only a small (if any) speed up is seen. Likely, the time spent executing each opcode overwhelms the
dispatch overhead. However, the work done in abstracting the generation of
the VM code will help in further experiments.